+++ Scraps.Next/Exceptions/AccountBannedException.cs
@@ -1,0 +1,8 @@
+﻿namespace Scraps.Next.Exceptions;
+
+public class AccountBannedException : Exception
+{
+    public AccountBannedException() {}
+    public AccountBannedException(string message) : base(message) {}
+    public AccountBannedException(string message, Exception inner) : base(message, inner) {}
+}